requirements.txt
selenium>=4.0.0
beautifulsoup4>=4.9.0
pandas>=1.3.0
requests>=2.26.0
lxml>=4.6.3
python-dotenv>=0.19.0
2captcha-python
